+++ Assets/Scripts/Projectile.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Projectile : MonoBehaviour
 {
-    private int damage = 1;
+    private float damage = 1;
     private Vector2 moveDirection;
     private float speed;
     private Rigidbody2D rb;
@@ -17,18 +17,15 @@ public class Projectile : MonoBehaviour
         moveDirection = direction;
         speed = launchSpeed;
 
-        // Задаем скорость Rigidbody
         if (rb != null)
         {
             rb.velocity = moveDirection * speed;
         }
 
-        // Разворачиваем спрайт шара по направлению полета (опционально)
         float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
         transform.rotation = Quaternion.AngleAxis(angle, Vector3.forward);
     }
 
-    // Нанесение урона при контакте (коллайдер шара должен быть IS TRIGGER)
     private void OnTriggerEnter2D(Collider2D collision)
     {
         if (collision.CompareTag("Player"))
@@ -39,12 +36,10 @@ public class Projectile : MonoBehaviour
                 playerHealth.TakeDamage(damage);
             }
 
-            // Уничтожаем шар после попадания в игрока
             Destroy(gameObject);
         }
     }
 
-    // АВТОУНИЧТОЖЕНИЕ: Срабатывает, когда шар полностью улетает из поля зрения ВСЕХ камер Unity
     private void OnBecameInvisible()
     {
         Destroy(gameObject);
